new commit

+++ src/main/java/assignment/training/service/ServiceImpl.java
@@ -1,9 +1,14 @@
 package assignment.training.service;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 
 
 
+import assignment.training.exception.PersonIsAlreadyPresent;
+import assignment.training.exception.UserNotFoundException;
 import assignment.training.model.Person;
 import assignment.training.repository.PersonDao;
 
@@ -14,28 +19,51 @@ public class ServiceImpl implements Service {
 	PersonDao dao;
 
 	@Override
-	public boolean insertPerson(Person person) {
+	public boolean insertPerson(Person person) throws PersonIsAlreadyPresent {
+		if(dao.findById(person.getId()).isPresent()) {
+			throw new PersonIsAlreadyPresent("person is already present");
+		}
 		dao.save(person);		
 		return true;
 	}
-	public Person updatePerson(Integer id, Person person) {
-		dao.save(person);
-		return dao.findById(id).get();
+	public Person updatePerson( Person person) throws UserNotFoundException {
+		Optional<Person> tempPerson=dao.findById(person.getId());
+		if(! tempPerson.isPresent()) {
+			throw new UserNotFoundException("person is not  present");
+		}
+		
+		
+		return dao.save(person);
 		
 	}
 	@Override
-	public void deletePerson(Integer id) {
+	public String deletePerson(Integer id) throws UserNotFoundException {
+		Optional<Person> tempPerson=dao.findById(id);
+		if(! tempPerson.isPresent()) {
+			throw new UserNotFoundException("person is not  present");
+		}
 		
 		 dao.deleteById(id);
+		 return "sucess";
 	}
 	
 	@Override
-	public boolean findPerson(Integer id) {
+	public boolean findPerson(Integer id)  {
 		if(dao.findById(id).isPresent()){
 			return true;
 		}
 		
 		return false;
+	}
+	@Override
+	public void deleteAll() {
+	dao.deleteAll();
+		
+	}
+	@Override
+	public List<Person> getAllPerson() {
+		return dao.findAll();
+	
 	}
 
 }
